Turn mst cost as long long int. As a consequence there were changes in mst.h/mst.c/main(s). Modify main for testing purposes to reflect MST/main

+++ FastMST/DataStructureOnGpu.h
@@ -20,7 +20,7 @@ struct DatastructuresOnGpu {
 	unsigned int* S;
 
 
-	unsigned int cost = 0;
+	unsigned long long int cost = 0;
 
 	void printForWebgraphvizrint() {
 		unsigned int	*e		= (unsigned int*)malloc(sizeof(unsigned int)* numEdges),
